ejercicio 21 Ejercicios de Repaso tema 5

+++ Repaso/Tema5/Ejercicio20/src/ejercicio20/Cuenta.java
@@ -45,7 +45,7 @@ public class Cuenta {
         this.saldo = saldo;
     }
     //Getters y Setters
-    public double getSaldo() {
+    public float getSaldo() {
         return saldo;
     }
